backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
bcrypt==4.0.1
python-multipart==0.0.13